datagen/requirements.txt
clickhouse-driver>=0.2.6,<0.3.0
psycopg2-binary>=2.9.9,<3.0.0
Faker>=24.0.0,<31.0.0
python-dotenv>=1.0.0,<2.0.0
